refactor(lib/oauth): move types to lib; brick re-exports
